feat: default to embedding for functions with cloudFront event
BREAKING CHANGE
Functions that are triggered by a cloudFront event embed the environment
variable statically by default. This is because these are lambda@edge
functions so will not have access to environment variables.

The boolean function.includeStaticEnv can override this behaviour. Set
to false on a function with a cloudFront trigger and environment will
not be embedded (just as if this plugin was not present). Set to true on
a function that wouldn't be enabled by default to embed the environment